resources_service/pyproject.toml
[tool.poetry]
name = "resources-service"
version = "0.1.0"
description = "Resources service for Travel Agent bot"
authors = ["lysmux <kllraz@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
pydantic-settings = "^2.2.1"
uvicorn = { extras = ["standart"], version = "^0.28.0" }
httpx = "^0.27.0"
geopy = "^2.4.1"
cashews = "^7.0.2"
redis = { extras = ["async"], version = "^5.0.3" }
faker = "^24.3.0"
pillow = "^10.2.0"


[tool.poetry.group.dev.dependencies]
pyright = "^1.1.354"
flake8 = "^7.0.0"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "src"